Simplified the logic behind the connection state function.
git-svn-id: svn+ssh://zeus.eng.uah.edu/home/work/cpe453/team8/cobra/trunk@5 7a5ebaa7-63b7-0410-984a-e5fb08a8dfaf

+++ src/event/auth.cc
@@ -99,7 +99,7 @@ cobraAuthEventHandler::handleServerEvent(cobraNetEvent* event)
     cobraStateEvent* newState = new cobraStateEvent();
     newState->setDestination(auth->source());
     newState->setSource(SERVER);
-    newState->setState(authorized ? QAbstractSocket::ConnectedState : QAbstractSocket::ConnectionRefusedError);
+    newState->setState((int)(authorized ? (int)QAbstractSocket::ConnectedState : (int)QAbstractSocket::ConnectionRefusedError));
 
     return true;
 }
